project-2-backend/requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.6
dnspython==2.4.2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-PyMongo==2.2.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
PyJWT==2.8.0
pymongo==3.13.0
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
Werkzeug==2.3.7
